version 1.5.6
Updated replication function to address issue with file permissions on
source catalog. Now using rsync with the proper options to reset the
permissions in place of scp. Also wrapped the Get-Content parsing of
the OVF in a Try/Catch to catch the failure due to lack of Read
permission on the transferred file.